requirements.txt
langchain-community
openai
gradio
chromadb
tiktoken